More Environment Assets and Texture Pool Boost
Ammended the DefaultEngine.ini to give 1Gb of texture streaming pool.  Begone error message!
Added basic mesh for 'Satellite Balloon' and several strut variants.
Edited the BP_Spline and its struct to allow scaling and use more sensible co-ordinates.
Added stripped down version of the city as City_Base and included a support for use on that.
Changed Support1 to have a top, in case of supporting small objects.
Added translucent and radar textures to use in new materials on new version of the bike.